stable version to work with app in google play

+++ src/controllers/Users/Follow.js
@@ -51,37 +51,45 @@ exports.unfollow = async (req, res) =>{
   const candidate = await knex('candidates')
     .select("id")
     .where('login', login)
-
-  if(type_user === 'user'){ //usuário normal
-    knex('followers').select('user_id').where('user_id', user_id).andWhere('candidate_id', candidate[0].id)
-      .then(userData =>{
-          if(userData.length !== 0){
-              return knex('followers')
-                      .select('*')
-                      .where('user_id', user_id)
-                      .delete()
-                      .then(() =>{
-                          res.status(200).json({msg:"UNFOLLOw"})
-                      })
-          }
-          return res.json({msg:"YOU ARE NOT FOLLOWING THIS CANDIDATE"});
-      })
-  }else{ //candidato
-    knex('followers').select('user_id').where('candidate', user_id).andWhere('candidate_id', candidate[0].id)
-      .then(userData =>{
-          if(userData.length !== 0){
-              return knex('followers')
-                      .select('*')
-                      .where('candidate', user_id)
-                      .delete()
-                      .then(() =>{
-                          res.status(200).json({msg:"UNFOLLOw"})
-                      })
-          }
-          return res.json({msg:"YOU ARE NOT FOLLOWING THIS CANDIDATE"});
-      })
+  if(candidate.length > 0){
+    if(type_user === 'user'){ //usuário normal
+      knex('followers')
+        .select('user_id')
+        .where('user_id', user_id)
+        .where('candidate_id', candidate[0].id)
+        .then(userData =>{
+            if(userData.length !== 0){
+                return knex('followers')
+                        .select('*')
+                        .where('user_id', user_id)
+                        .where('candidate_id', candidate[0].id)
+                        .delete()
+                        .then(() =>{
+                            res.status(200).json({msg:"UNFOLLOw"})
+                        })
+            }
+            return res.json({msg:"YOU ARE NOT FOLLOWING THIS CANDIDATE"});
+        })
+    }else{ //candidato
+      knex('followers').select('user_id').where('candidate', user_id).andWhere('candidate_id', candidate[0].id)
+        .then(userData =>{
+            if(userData.length !== 0){
+                return knex('followers')
+                        .select('*')
+                        .where('candidate', user_id)
+                        .where('candidate_id', candidate[0].id)
+                        .delete()
+                        .then(() =>{
+                            res.status(200).json({msg:"UNFOLLOw"})
+                        })
+            }
+            return res.json({msg:"YOU ARE NOT FOLLOWING THIS CANDIDATE"});
+        })
+    }
+  }else{
+    return res.json({msg:"YOU ARE NOT FOLLOWING THIS CANDIDATE"});
   }
-    
+  
 }
 
 // PEGAR SEGUIDORES DE UM CANDIDATO ESPECIFICO
@@ -102,14 +110,18 @@ exports.getFollowersCount = async (req, res) => {
   const candidate = await knex('candidates')
     .select("id")
     .where('login', login)
+  if(candidate.length > 0){
+    knex('followers').select('*').where('candidate_id', candidate[0].id)
+      .then(result => {    
+          res.status(200).json(result.length)
+      }).catch((error) => {
+        console.log(error)
+          res.status(404).json({msg:"YOU DONT HAVE FOLLOWERS"})
+      })
+  }else{
+    res.status(404).json({msg:"YOU DONT HAVE FOLLOWERS"})
+  }
   
-  knex('followers').select('*').where('candidate_id', candidate[0].id)
-  .then(result => {    
-      res.status(200).json(result.length)
-  }).catch((error) => {
-    console.log(error)
-      res.status(404).json({msg:"YOU DONT HAVE FOLLOWERS"})
-  })
 }
 
 exports.isFollower = async (req, res) => {
@@ -156,7 +168,6 @@ exports.isFollower = async (req, res) => {
 exports.getFollowed = async (req, res) => {
   const user_id = req.params.user_id
   const type_user = req.params.type_user //tipo do usuário que está fazendo a consulta, candidate ou user
-
   
   if(type_user === 'user'){ //consulta normal
     knex('candidates')
@@ -175,20 +186,6 @@ exports.getFollowed = async (req, res) => {
         res.status(400).json({msg:"Você ainda não segue ninguém"})
       })
   }else{ //caso seja candidato
-    // knex('followers')
-    //   .select('candidate_id') //pega só o id do candidato que a pessoa segue
-    //   .where('candidate', user_id) //pega na coluna do id do candidato que segue outro candidato
-    //   .orderBy('id', 'desc')
-    //   .then(result => {    
-    //     if(result.length > 0){
-    //       res.status(200).json(result)  
-    //     }else{
-    //       res.status(400).json({msg:"Você ainda não segue ninguém"})
-    //     }
-    //   }).catch((error) => {
-    //     //console.log(error)
-    //       res.status(400).json({msg:"Você ainda não segue ninguém"})
-    //   })
     knex('candidates')
       .select('candidates.id','name','party','coalition','description','city_id','state_id','number','profile_pic','cover_pic','badges','proposals','login') 
       .join('followers', 'followers.candidate_id', 'candidates.id')
